About
Added about.fxml scene which will display info on myself and the project. Added the Library.jar as well as the shell script to run the application on my computer by double clicking. Added about button to main menu as well as updated the controllers to swap between about and main menu.