changed a few things on main nbd
Added two actual books and found out I needed to account for the description to contain commas. Implemented some regex I found online that I can't read but it works so I'mmmmmmma leave it. Changed book isbn back to a long because it needs to be end of story no questions good bye.